Added course dedicated images mechanism
This commit contains a patch that enables the module Certificate
to have specific borders,seals,signatures and watermarks visible
only inside a specific courses.

Inside every pix directory there is a new directory named
by_course_shortname which should contain only the images that
are dedicated to specific courses. The naming convention for these
images is COURSE_SHORTNAME_description.extension.

So for example:
	pix/borders/by_course_shortname/course_007_purple_border.jpg

The border purple_border is going to be visibile only inside
the certificates of the course that has course_007 as shortname.
For every other course this border is going to be invisible.

This applies to every other directory inside pix directory.

Please note that every image inside by_course_shortname directory
that doesn't follow the naming convention is going to be ignored.