BUGFIX: Filename length check
The filename length check was missing.
The length of the filenames is constrained by the table
definitions inside the module table.

This commit implements a checking mecanism.

The costraint is 30 chars for general images while for the
images in the "c" folders the max filename length is 28 chars.

The filename length limitation is releated to the tables of this module.

The filename length check is enforced by the FILE_NAME_LENGTH constant
in the lib.php file. Please change this costant if you modify
the certificate table definition.

This commit adds FILE_NAME_LENGTH check not only for the course
specific images but also for default images. This means that the files
longer than 30 chars inside folders: borders, watermarks, seals and
signatures are going to be ignored.

Without this check all files with the names longer than 30 chars
used to produce errors when chosen.