Implemented Course protected images mechanism
The Course protected images allows the user to upload images
that are visible only from Course specified during the upload.

The global uploading form has been extended by adding two fields:
- Limit to one course (Checkbox)
  Selecting this checkbox is going to place the upladed image inside
  the $CFG->data_root/mod/certificate/image_type/course/$COURSE->shortname
  folder. The images are then going to be visible only form the course
  with the same shortname.

- Course Shortname
  This field specifies the Course shortname to use during the upload.
  Thus the file is going to be placed in the directory specified by
  the input of this field. Before proceeding with the update the module
  validates the input by checking the existence of the course.

This is particularly useful for the Signature images because
it avoids the signature to be visible in all other courses once they
are uploaded.

This commits also adds the Italian language support.